# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ZeroTrace API
  description: |
    ZeroTrace API provides comprehensive security scanning and vulnerability management capabilities.
    The API supports real-time scanning, compliance monitoring, and advanced analytics.
  version: 2.0.0
  contact:
    name: ZeroTrace Support
    email: support@zerotrace.io
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.zerotrace.io
    description: Production server
...
paths:
...
  /api/agents/register:
    post:
      tags:
        - Agents
      summary: Register agent
      description: Register a new agent with the system
      operationId: registerAgent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentRegistration'
      responses:
        '200':
          description: Agent registered successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/agents/heartbeat:
    post:
      tags:
        - Agents
      summary: Agent heartbeat
      description: Update agent heartbeat status
      operationId: agentHeartbeat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentHeartbeat'
      responses:
        '200':
          description: Heartbeat updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
  /api/agents/results:
    post:
      tags:
        - Agents
      summary: Submit scan results
      description: Submit vulnerability scan results from agent
      operationId: agentResults
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AgentResults'
      responses:
        '200':
          description: Results submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
...
components:
  schemas:
    APIResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        error:
          $ref: '#/components/schemas/APIError'
    APIError:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: string
    Agent:
      type: object
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        status:
          type: string
          enum: [active, inactive, offline]
        version:
          type: string
        hostname:
          type: string
        os:
          type: string
        last_seen:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
    AgentRegistration:
      type: object
      required:
        - id
        - organization_id
        - name
        - version
        - hostname
        - os
      properties:
        id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        name:
          type: string
        version:
          type: string
        hostname:
          type: string
        os:
          type: string
    AgentHeartbeat:
      type: object
      required:
        - agent_id
        - organization_id
        - agent_name
        - status
      properties:
        agent_id:
          type: string
          format: uuid
        organization_id:
          type: string
          format: uuid
        agent_name:
          type: string
        status:
          type: string
        cpu_usage:
          type: number
          format: float
        memory_usage:
          type: number
          format: float
        metadata:
          type: object
        timestamp:
          type: string
          format: date-time
    AgentResults:
      type: object
      required:
        - agent_id
        - results
      properties:
        agent_id:
          type: string
          format: uuid
        results:
          type: array
          items:
            $ref: '#/components/schemas/AgentScanResult'
        metadata:
          type: object
    AgentScanResult:
      type: object
      properties:
        scan_type:
          type: string
        dependencies:
          type: array
          items:
            $ref: '#/components/schemas/Dependency'
        vulnerabilities:
          type: array
          items:
            $ref: '#/components/schemas/Vulnerability'
        timestamp:
          type: string
          format: date-time
    Dependency:
      type: object
      properties:
        name:
          type: string
        version:
          type: string
        type:
          type: string
        path:
          type: string
    Vulnerability:
      type: object
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        severity:
          type: string
          enum: [critical, high, medium, low]
        category:
          type: string
        cve_id:
          type: string
        cvss_score:
          type: number
          format: float
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/APIError'
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token authentication
security:
  - BearerAuth: []